+++ plone/app/layout/viewlets/tests/test_content.py
@@ -32,8 +32,8 @@ class TestDocumentBylineViewletView(ViewletsTestCase):
     Test the document by line viewlet
     """
     def afterSetUp(self):
-        addMember(self, 'Alan', roles=('Member', 'Manager'))
-        addMember(self, 'Ano', roles=())
+        addMember(self.portal, 'Alan', roles=('Member', 'Manager'))
+        addMember(self.portal, 'Ano', roles=())
 
     def test_anonymous_locked_icon(self):
         request = self.app.REQUEST
@@ -60,13 +60,13 @@ class TestDocumentBylineViewletView(ViewletsTestCase):
         lockIconUrl = '<img src="http://nohost/plone/lock_icon.png" alt="" \
 title="Locked" height="16" width="16" />'
         self.assertEqual(viewlet.locked_icon(), lockIconUrl)
-    
+
     def test_pub_date(self):
         request = self.app.REQUEST
         self.login('Alan')
         self.portal.invokeFactory('Document', 'd1')
         context = getattr(self.portal, 'd1')
-        
+
         # configure our portal to enable publication date on pages globally on
         # the site
         properties = getToolByName(context, 'portal_properties')
@@ -76,7 +76,7 @@ title="Locked" height="16" width="16" />'
         self.login('Ano')
         viewlet = DocumentBylineViewlet(context, request, None, None)
         viewlet.update()
-        
+
         # publication date should be None as there is not Effective date set for
         # our document yet
         self.assertEqual(viewlet.pub_date(), None)
@@ -85,7 +85,7 @@ title="Locked" height="16" width="16" />'
         effective = DateTime()
         context.setEffectiveDate(effective)
         self.assertEqual(viewlet.pub_date(), DateTime(effective.ISO8601()))
-        
+
         # now switch off publication date globally on the site and see if
         # viewlet returns None for publication date
         site_properties.displayPublicationDateInByline = False